Add auto-scrolling; Right-click menu with settings
Now three settings can be toggled through the right-click menu―
whether or not the view should be transparent when no lyrics are
available (or no music is playing); whether or not the dragger should
be, as well; and whether or not auto-scroll should be toggled.

Auto-scroll, when enabled, will scroll through the text automatically,
in proportion with current position in the song.